[DRUP-388] Test the new version of the authentication form.
